8.4 Fusionner deux classes
